perf(workbin table): abstract row components and add memo and improve table name sorting
